Update README.md
About Dataset
Introduction
Financial loan services are leveraged by companies across many industries, from big banks to financial institutions to government loans. One of the primary objectives of companies with financial loan services is to decrease payment defaults and ensure that individuals are paying back their loans as expected. In order to do this efficiently and systematically, many companies employ machine learning to predict which individuals are at the highest risk of defaulting on their loans, so that proper interventions can be effectively deployed to the right audience.
Description
This dataset has been taken from Coursera's Loan Default Prediction Challenge and will provide you the opportunity to tackle one of the most industry-relevant machine learning problems with a unique dataset that will put your modeling skills to the test. The dataset contains 255,347 rows and 18 columns in total.
 





New Skills
I want to deepen my understanding of dimensionality reduction: feature selection and feature removal.
Build a classification model to predict clients who are likely to default on their loan and give recommendations to the bank on the important features to consider while approving a loan.
Some of the questions that I want to explore include how can bank identify the defaulter who might missed payments or totally non-payment of their loan and reduce the losses for the company caused by defaulter?
I also want to further strengthen my skills on improving model performance.
Use regularization to prevent overfitting. Regularization is a technique that adds a penalty to the model's loss function, which helps prevent the model from overfitting the training data. There are different types of regularization, such as L1 and L2 regularization, which add different types of penalties to the loss function.

I will follow this process:
1.	Using Loan Default Prediction
2.	Cleaning and Preprocessing this Data - removing errors, filling gaps, and adjusting the data format.
3.	Selecting the right model- Since in previous modules I had applied other models, I will try to explore neural networks. Dimensionality reduction: feature selection and feature removal.
4.	Tuning hyperparameters - test different values—e.g., try 2 layers vs. 5 layers. If your model predicts too many non-defaulters as defaulters, tweaking the learning rate might help it learn more accurately.
5.	Improving model performance- Making the model better at predicting loan defaults by fixing its weaknesses and boosting its accuracy